Tried to make as many thing predicted as possible
