+++ Content.Client/Species/CocoonVisualizerSystem.cs
@@ -26,7 +26,6 @@ public sealed class CocoonVisualizerSystem : EntitySystem
         base.Initialize();
 
         SubscribeNetworkEvent<CocoonRotationAnimationEvent>(OnCocoonRotationAnimation);
-        SubscribeLocalEvent<CocoonedComponent, AttackAttemptEvent>(OnCocoonedAttackAttempt);
     }
 
     private void OnCocoonRotationAnimation(CocoonRotationAnimationEvent args)
@@ -118,11 +117,5 @@ public sealed class CocoonVisualizerSystem : EntitySystem
         };
 
         _animation.Play((uid, animationComp), animation, animationKey);
-    }
-
-    private void OnCocoonedAttackAttempt(Entity<CocoonedComponent> ent, ref AttackAttemptEvent args)
-    {
-        // This prevents the client-side attack animation from playing
-        args.Cancel();
     }
 }
