+++ Content.Client/Species/Systems/CocoonSystem.cs
@@ -4,6 +4,19 @@ namespace Content.Client.Species.Arachnid;
 
 public sealed class CocoonSystem : SharedCocoonSystem
 {
+    protected override void OnWrapDoAfterServer(EntityUid performer, EntityUid target, EntityUid cocoonContainer)
+    {
+    }
 
+    protected override void OnWrapActionServer(EntityUid user, EntityUid target)
+    {
+    }
+
+    protected override void OnWrapDoAfterSetupVictimEffects(EntityUid victim)
+    {
+    }
+
+    protected override void OnCocoonContainerShutdownRemoveMumbleAccent(EntityUid victim)
+    {
+    }
 }
-
