+++ Resources/Locale/en-US/actions/actions/spider.ftl
@@ -12,6 +12,8 @@ arachnid-wrap-start-target = {$user} begins wrapping you in sticky silk!
 arachnid-wrap-complete-user = You bind {$target} inside a cocoon of silk.
 arachnid-wrap-complete-target = Sticky silk envelops you, leaving you helpless!
 arachnid-unwrap-self = You're trussed up too tightly to free yourself!
+arachnid-break-free-start = You begin struggling to break free from the cocoon...
+arachnid-break-free-complete = You manage to break free from the cocoon!
 arachnid-unwrap-start-user = You begin unwrapping the cocoon restraining {$target}.
 arachnid-unwrap-start-user-empty = You begin unwrapping the cocoon.
 arachnid-unwrap-start-target = {$user} begins to unwrap you from the cocoon!
