+++ global.json
@@ -1,6 +1,7 @@
 {
   "sdk": {
     "version": "10.0.100",
-    "rollForward": "latestFeature"
+    "rollForward": "latestFeature",
+    "allowPrerelease": true
   }
 }
